Implemented CollectPerBaseCounts
CollectPerBaseCounts falls under the umbrella of allelic analysis along
with CollectAllelicCounts. The important differences between these tools
are that CollectPerBaseCounts reports the counts of all nucleotides {A,
C, G, T, N} and only reports a maximum of one count per template.
CollectAllelicCounts, on the other hand, reports only the two most
commons nucleotides and counts bases from both reads of the same
fragment.

The overall implementation and testing is very similar to
CollectAllelicCounts.