Changes to help to debug the (local) Java VM using the standard agent.
The ./gatk script has a couple o new arguments --java-debug and --java-debug-suspend. Both optionally take on a
numerical argument (only positive integers are supported) that would indicate the localhost port the agent is going to listed from

--java-debug <port>? would start the debug agent without suspending the VM whereas --java-debug-suspend <port>? suspends the
VM. If both are provided --java-debug-suspend takes preference.

I allow to speficify a listening port (5005) by default and instruct the VM to suspend and wait for the debugger
to connect.