Move to GenomicsDB 1.3.0 and add support for shared posixfs optimizations (#6654)
* Move to GenomicsDB Version 1.3.0 for optimization options, pre-compression filters and bug fixes

Add new arg --genomicsdb-shared-posixfs-optimizations to help with shared posix filesystems like NFS and Lustre. This turns on disable file locking and for GenomicsDBImport it minimizes writes to disks. The performance on some of the gatk datasets for the import of about 10 samples went from 23.72m to 6.34m on NFS which was comparable to importing to a local filesystem. 

This version of GenomicsDB also uses pre-compression filters for offset and compression files for new workspaces and genomicsdb arrays. The total sizes for a GenomicsDB workspace using the same dataset as above and the 10 samples went from 313MB to 170MB with no change in import and query times. Smaller GenomicsDB arrays also help with performance on distributed and cloud file systems.

This version has added support to handle MNVs similar to deletions as described in Issue 6500. See GenomicsDB PR 88. 

There is added support in the GenomicsDBImporter to have multiple contigs in the same GenomicsDB partition/array. This will hopefully help import times in cases where users have many thousands of contigs. See GenomicsDB PR 91. Changes will be needed from the gatk side to leverage this support. 

Logging has been improved somewhat with the native C/C++ code using spdlog and fmt and the Java layer using apache log4j and log4j.properties provided by the application. Also, info messages like No valid combination operation found for INFO field AA - the field will NOT be part of INFO fields in the generated VCF records will only be output once for the operation. Also see open PR 6514 for code to actually suppress these warnings for known fields.