Update the gcloud package signing key during the docker build
Previously we were relying on the gcloud package signing key retrieved
during build of the GATK base image. However, the base image is updated
so infrequently that it's possible that the key it uses has expired. To
address this, we now retrieve an updated key at the start of the Docker
build.