Cherry pick of commits to fix GATK tests from master (#7183)
* Disabling some spark dataproc tests because of dependency issues. (#7170)

Will re-enable in the near future.

* Update the gcloud package signing key during the docker build (#7180)

Previously we were relying on the gcloud package signing key retrieved
during build of the GATK base image. However, the base image is updated
so infrequently that it's possible that the key it uses has expired. To
address this, we now retrieve an updated key at the start of the Docker
build.

* fix compilation problem

Co-authored-by: Jonn Smith <jonn-smith@users.noreply.github.com>
Co-authored-by: droazen <droazen@broadinstitute.org>