GenomicsDBImport: add ability to specify explicit index locations via the sample name map file
The sample name map file accepted by GenomicsDBImport can now optionally contain a third
column giving an explicit path to an index for the corresponding VCF. It is allowed to
specify an explicit index in some lines of the sample name map and not others.

Added comprehensive unit and integration tests.